Added support for Weight value on nodes
Modified the slurm.conf template to include the Weight parameter if the
slurm-controller charm has been instructed to put weight on nodes via
the node_weight_criteria charm option.

When it is preferable to allocate for example smaller memory nodes for
smaller jobs, low weights should be assigned to smaller nodes.
Configuring node_weight_criteria on the slurm-controller charm  will
automatically order and weigh the nodes in ascending order. Allowed
values are RealMemory, CPUs and CoresPerSocket.